[Add]pinia
